server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["NikPrav <nikhil.pravin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "7.1.2"
Flask = "1.1.2"
itsdangerous = "1.1.0"
Jinja2 = "2.11.3"
MarkupSafe = "1.1.1"
numpy = "1.20.1"
Pillow = "8.1.2"
torch = "1.8.0"
torchvision = "0.9.0"
typing-extensions = "3.7.4.3"
Werkzeug = "1.0.1"
opencv-python = "^4.5.1"
mongoengine = "^0.23.0"
matplotlib = "^3.3.4"
gunicorn = "^20.0.4"
pydantic = {extras = ["dotenv"], version = "^1.8.1"}
odmantic = "^0.3.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
